fix: rename http.py to http_client.py to avoid Python stdlib conflict
The module name `http.py` conflicts with Python's standard library
`http` package. When importing `urllib.request`, Python finds our
`http.py` instead of the stdlib's `http.client` module.

Changes:
- Renamed http.py to http_client.py
- Updated all imports from `http` to `http_client`
- Added __init__.py to make lib a proper Python package
- Removed unused `import time` from download.py
