+++ turbo/apps/web/src/lib/e2b/scripts/lib/__init__.py.ts
@@ -1,0 +1,11 @@
+/**
+ * Python package init file for lib directory
+ * Makes the lib directory a proper Python package
+ */
+export const INIT_SCRIPT = `#!/usr/bin/env python3
+"""
+VM0 Agent Scripts Library
+
+This package contains utility modules for the VM0 agent execution in E2B sandbox.
+"""
+`;
